Fixed the null variable issue in honorMember call.
I fixed the null variable issue when calling the honorMember function in popup.js by moving the login function calls to the beginning of popup.js.
